--- java/standardization/README.md
@@ -1,2 +1,0 @@
-# algorithmia-java-pipe
-A package enabling deployment of Java algorithms on Algorithmia.
